refactor(table): accept pr suggestions
