+++ packages/admin-ui/src/table/index.tsx
@@ -1,6 +1,6 @@
-export * from './typings'
+export * from './types'
 export * from './components/table'
-export * from './createColumns'
+export * from './create-columns'
 export * from './components/table-head'
 export * from './components/table-body'
 export * from './components/table-cell'
